% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTasmapLayers.R
\name{listTasmapLayers}
\alias{listTasmapLayers}
\title{List Tasmanian maps layers}
\usage{
listTasmapLayers()
}
\value{
names vector
}
\description{
List Tasmanian maps layers
}
\examples{
listTasmapLayers()
}
