% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasmap.R
\name{tasmap}
\alias{tasmap}
\title{Tasmania maps}
\usage{
tasmap(map, data, ...)
}
\arguments{
\item{map}{optional mapview object}

\item{data}{see \code{mapview::mapview}}

\item{...}{see mapview}
}
\description{
Tasmania maps
}
\references{
https://atriplex.info/blog/index.php/2017/06/08/adding-wms-basemaps-to-a-mapview-map/
}
