% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasmapr-package.R
\docType{package}
\name{tasmapr-package}
\alias{tasmapr}
\alias{tasmapr-package}
\title{tasmapr: Tasmania Mapserver}
\description{
Tile map servers from 'TheList' <https://maps.thelist.tas.gov.au/listmap/app/list/map>.
}
\author{
\strong{Maintainer}: Michael Sumner \email{first.last@example.com}

Authors:
\itemize{
  \item Grant Williamson \email{first.last@example.com}
}

}
\keyword{internal}
