% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasmap.R
\name{tasmap}
\alias{tasmap}
\title{Tasmania maps}
\usage{
tasmap(map, data, basemap = "Topographic", ...)
}
\arguments{
\item{map}{optional mapview object}

\item{data}{see \code{mapview::mapview}}

\item{basemap}{short name of Tasmap baselayer, see \code{tasmapr::listTasmapLayers} or the URL of the WMS service to use}

\item{...}{see mapview}
}
\value{
mapview object
}
\description{
Tasmania maps
}
\examples{
tasmap(basemap = "TTSA")
}
\references{
https://atriplex.info/blog/index.php/2017/06/08/adding-wms-basemaps-to-a-mapview-map/
}
